requirements.txt
absl-py==0.14.0
astunparse==1.6.3
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-bootstrap-templates==0.1.1
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.2
plotly==5.2.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
